+++ src/main/java/com/initsan/myToDoList/repository/TasksRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TasksRepository extends JpaRepository<Task, Integer> {
+public interface TasksRepository extends JpaRepository<Task, Long> {
     Optional<Task> findByTitle(String title);
 }
